A prototype of two related optimizations: basic block simplification, and recognition of zeroing loops.
Basic block simplication transforms straight-line sequences of
operations into a simpler form. Adjustments of the tape head are
merged into a single adjustment at the end of the block. Modifications
to the tape are merged into a single add-at-offset / subtract-at-offset
operation per tape offset that performs the aggregate of all
modifications to that location within the block. This allows sequences
such as `>>+<<` to be converted to `AddAtOffset(1, 1)`, and `->>+>>` to be
represented as `Subtract(1), AddtOffset(1, 2), MoveRight(4)`.

Recognition of zeroing loops detects the `[-]` pattern that loops until
the current tape location is zero. These loops are instead replaced with
an explicit operation to set the current tape location to zero. When
combined with basic block simplification, this allows sequences such as
`[-]++` to be represented as `SetAtOffset(0, 0)`, with the resulting
benefits in generated machine code size.

Loop detection should be tackled more generally. Loop bodies that in
aggregate do not advance the tape head, such as `[>+<-]`, should be
recognized and flattend into higher-level operations such as
`AddFromOffsetToOffset(0, 1), SetAtOffset(0, 0)`. This wil reveal
further opportunities for basic block simplification. Such optimizations
will likely be easier if the intermediate representation is changed to
represent loops as a single operation with their bodies as children,
rather than unrelated pairs of instructions.

This change also includes the assembler modifications necessary to
support `MovIM` and signed offsets in instructions that reference memory
locations. `MovIM` needs further testing, while the signed offset stuff
needs rethought since it flowed further through the assembler than was
appropriate. In particular, it made its way down to the
`emit_n_bit_constant` functions that are also used for unsigned data
such as immediates.