Add an optimization pass to simplify mutation sequences.
Simplication of mutation sequences transforms straight-line sequences of
operations into a simpler form. Adjustments of the tape head are merged
into a single adjustment at the end of the block. Modifications to the
tape are merged into a single add / subtract / set operation per tape
offset that performs the aggregate of all modifications to that location
within the block. This allows sequences such as `>>+<<` to be converted
to `Add(1, 1)`, and `->>+>>` to be represented as
`Subtract(1), Add(1, 2), MoveRight(4)`.