Add a work-in-progress JIT based on LLVM.
Code generation is completely implemented and generates correct results when
assembling the resulting bitcode file into a command-line executable. At
this point it does not look like it will be a performance improvement
over the custom JIT, even for programs as complicated as Mandelbrot, as
code generation and optimization are relatively expensive. It's likely
that the LLVM JIT would be an improvement for programs running for tens
of seconds.

The JIT is currently non-functional as LLVM appears to incorrectly link to
external functions used by the generated code. Calls to `memset` and
`write` end up jumping to unmapped memory rather than the expected
functions.

To enable LLVM support, pass `--features llvm` to `cargo`.