Fix system calls on macOS
System call numbers are grouped into different classes on x86_64 macOS, with the class being represented in the high
bits of the system call number. We need to OR SYSCALL_CLASS_UNIX into our system call numbers so that macOS knows
which class to interpret the system call as being from.